\name{GOstats-defunct}
\alias{GOstats-defunct}
\title{Defunct Functions in GOstats Package}
%------ PLEASE: put \alias{.} here for  EACH !
\alias{combGOGraph}
\description{
  The functions or variables listed here are no longer part of GOstats as
  they are not needed (any more).
}
\usage{
combGOGraph(g1, g2)
}
\arguments{
  \item{g1}{A graph}
  \item{g2}{A graph}
  }
\details{
  \code{combGOGraph} was replaced by \code{join}.
}

\keyword{manip}
